kerberos: Style fixes
Let's follow the same error handling idiom as used in the rest of the
code.

It is good to use a GDestroyNotify with g_task_return_pointer because
the pointer won't be leaked if g_task_propagate_pointer is not called.
This brings the code more in line with what is expected from a
GAsyncResult-based API.